refactor(mobile): decompose CalendarView; wire mobile logic tests
CalendarView.tsx had grown to ~930 lines mixing grid/date math, the host
component, and five presentational sub-components under one big makeStyles.
Split it (no behaviour or styling change; comments preserved verbatim):

- components/calendar/grid.ts (+12 unit tests) — Cell + dayKey/dueDayKey/
  buildMonthGrid/buildWeekGrid/formatDayLabel. Kept date-fns-free and
  mobile-local; deliberately NOT shared with the desktop calendar utils
  (which use date-fns + a different cell shape).
- components/calendar/styles.ts — makeCalendarStyles moved verbatim +
  useCalendarStyles hook + CalendarStyles type. Host computes styles once and
  passes them down as a prop, unchanged perf model.
- components/calendar/{TaskChip,DraggableChip,DayCell,DayDetailRow,
  UnscheduledTray}.tsx — the presentational pieces.

Host CalendarView.tsx (934 -> 352 lines) is now state + drag wiring +
bucketing memo + composition. Public exports (CalendarView, CalendarLayout)
kept at the same path, so both callers are untouched.

Also adds a 'mobile' vitest project (node env, @->mobile/src alias,
mobile/**/*.test.ts) so mobile pure-logic tests run in the shared runner for
the first time — which also revives the previously-unrun
src/dnd/autoScroll.test.ts (14 tests). Full suite: 1128 passing.
