+++ mobile/changelogs/v0.1.3.md
@@ -64,3 +64,4 @@ Polish for the workspace switcher and Knowledge Graph controls.
 - Note and task semantic search now share one ranking pipeline (`rankSemantic`) parameterised by row source and id→text/title lookups, instead of two near-identical copies — so query-cache, centroid centring, hybrid dense+lexical rerank and slicing can't drift apart between notes and tasks. No behaviour change.
 - Syntax-highlighting grammars (lowlight/highlight.js `common` set) now construct lazily on the first code block render instead of at module import, so app launch and the first notes/chat render don't pay the grammar-registration cost when no code is shown.
 - Split the Chat screen — which had grown to one 800-line file mixing message rendering, the composer, keyboard/scroll choreography and tool-navigation logic — into focused pieces (a message bubble, reasoning disclosure, tool-chip trail, composer, a keyboard-scroll hook, and shared/unit-tested tool-navigation logic), and made it load its saved history only once on open instead of twice. No visible change; the chat behaves and looks exactly the same, but the code is far easier to change safely.
+- Split the Calendar view (previously a single ~930-line file) into focused pieces — the month/week grid math, the day cell, task chips, the unscheduled tray, and the day-detail row — with the grid/date logic now covered by unit tests. Mobile logic tests are also wired into the shared test runner for the first time (the calendar date math and the drag auto-scroll helpers now run in CI). No visible change; the calendar looks and behaves exactly the same.
